Move container image handling to the provisioning package
