+++ provisioning/sample_images.json
